Rename packet to info (refactoring)
